+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 
 ### Libraries
 
+#### `dart:async`
+
+- Added `broadcast` parameter to `Stream.empty` constructor.
+
 #### `dart:developer`
 
 - Deprecated the `Service.getIsolateID` method.
@@ -30,7 +34,7 @@
   `HttpClientResponse.headers` and `HttpRequest.headers` no longer include
   trailing whitespace in their values.
 
-[#52334]: https://dartbug.com/53005
+[#53005]: https://dartbug.com/53005
 
 #### `dart:js_interop`